fastapideta/requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
PyMySQL
passlib
bcrypt
python-jose
python-multipart